Moved classes into separate files.
