+++ lib/commands/place.rb
@@ -1,0 +1,6 @@
+class Place
+
+  def self.execute(args)
+    toy_robot = ToyRobot.new(args)
+  end
+end
