+++ lib/commands/report.rb
@@ -1,0 +1,6 @@
+class Report
+
+  def self.execute(toy_robot)
+    puts [toy_robot.x, toy_robot.y, toy_robot.facing]
+  end
+end
