+++ lib/facing_instructions.rb
@@ -1,0 +1,11 @@
+module FacingInstructions
+
+  def facing_instructions
+    {
+      'NORTH' => 'EAST',
+      'EAST' => 'SOUTH',
+      'SOUTH' => 'WEST',
+      'WEST' => 'NORTH',
+    }
+  end
+end
